Rename `cats_effects` to `cats_effect`
