+++ src/main/scala/com/github/gvolpe/cats_effect/Once.scala
@@ -1,4 +1,4 @@
-package com.github.gvolpe.cats_effects
+package com.github.gvolpe.cats_effect
 
 import cats.effect._
 import cats.effect.concurrent.Deferred
